% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_maker.R
\name{gen_mix_dist_maker}
\alias{gen_mix_dist_maker}
\alias{samemix_dist_maker}
\title{Generates a general mixture distribution}
\usage{
gen_mix_dist_maker(
  type = c("same", "general"),
  families = c("normal"),
  nr_distributions = 1L,
  trafos_each_param = NULL,
  inflation_values = NULL
)

samemix_dist_maker(family = "normal", nr_comps = 3, trafos_each_param = NULL)
}
\arguments{
\item{families}{character (vector); 
see the \code{family} argument of \code{deepregression}. Can be multiple
distributions for a mixture of different distributions}

\item{nr_distributions}{integer; defining the number of distributions
to be mixed (including deterministic ones)}

\item{trafos_each_param}{list of transformations applied before plugging
the linear predictor into the parameters of the distributions.
Should be of length #parameters of \code{dist}}

\item{family}{character; see the \code{family} argument of \code{deepregression}}

\item{nr_comps}{number of mixture components}

\item{dist_fun}{tfd distribution defining the mixture}
}
\value{
returns function than can be used as argument \code{dist\_fun} for
}
\description{
This function supplies the functionality to define mixture distribution
from TFP, including the
}
\examples{

library(deepregression)

# a function that can be passed to deepregression
# as family
gen_mix_dist_maker(
  mix_dist_fun = tfd_mixture_same_family,
  nr_distributions = 2L
 )
 
gen_mix_dist_maker(
  families = c("normal", "student_t"),
  trafos_each_param = 
  list(
       list(function(x) x, 
            function(x) tf$add(add_const, tfe(x))
       ),
  ),
  list(
       function(x) x, 
       function(x) tf$add(add_const, tfe(x))
       )
  ),
  mix_dist_fun = tfd_mixture,
  nr_distributions = 2L
 )

}
