% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.mixdistreg}
\alias{coef.mixdistreg}
\alias{plot.mixdistreg}
\title{Generic methods for mixdistreg objects}
\usage{
\method{coef}{mixdistreg}(object, which_dist = "mixture", which_param = 1L, type = NULL, ...)

\method{plot}{mixdistreg}(
  x,
  which_dist = "mixture",
  which = NULL,
  which_param = 1,
  only_data = FALSE,
  grid_length = 40,
  main_multiple = NULL,
  type = "b",
  get_weight_fun = get_weight_by_name,
  ...
)
}
\arguments{
\item{object}{mixdistreg object}

\item{which_dist}{integer or string. For integer, see
\code{?coef.deepregression}. If a string is provided, it
must be one of \code{"mixture"} or a family string used
to define the distributions in the mixture (e.g. \code{"normal"}).
Can also be a combination of the family and a number to specify which
of several distributions with the same family (e.g., \code{"normal2"}
for the second normal distribution)}

\item{which_param}{see \code{?coef.deepregression}}

\item{type}{see \code{?coef.deepregression}}

\item{...}{further arguments, passed to fit, plot or predict function}
}
\description{
The main difference to methods from deepregression
is that these allow to define strings for \code{which_param}
to explicitly extract a coefficients, plots, ... for 
either the \code{"mixture"} or a defined distribution family.
}
