% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_mix.R
\name{mixdistreg}
\alias{mixdistreg}
\title{Generic mixture distributional regression}
\usage{
mixdistreg(
  y,
  families = c("normal"),
  type = c("same", "general"),
  nr_comps = 2L,
  list_of_formulas,
  formula_mixture = ~1,
  list_of_deep_models = NULL,
  inflation_values = NULL,
  data,
  ...
)
}
\arguments{
\item{y}{response}

\item{families}{character (vector); 
see the \code{family} argument of \code{deepregression}. Can be multiple
distributions for a mixture of different distributions (\code{type = "general"})}

\item{nr_comps}{integer; number of mixture components}

\item{list_of_formulas}{a list of formulas for the 
parameters of the distribution(s) which is/are used in the mixture
(i.e., for \code{family="normal"} only a list of two formulas
is required). See also \code{?deepregression} for more details.}

\item{formula_mixture}{formula for the the additive predictor 
of the mixture component. Covariate-independent per default.}

\item{list_of_deep_models}{see \code{?deepregression}}

\item{data}{data.frame or list with data}

\item{...}{further arguments passed to \code{?deepregression}}
}
\value{
a model of class \code{mixdistreg} and
\code{deepregression}
}
\description{
Fit a mixture distributional regression
}
\examples{

}
